fix: title should not use 'in' when no ou selected
